Move request related repositories to a separate package CIRC-849
