+++ src/main/java/org/folio/circulation/domain/CreateRequestRepositories.java
@@ -1,9 +1,9 @@
 package org.folio.circulation.domain;
 
-import org.folio.circulation.infrastructure.storage.RequestPolicyRepository;
+import org.folio.circulation.infrastructure.storage.requests.RequestPolicyRepository;
 import org.folio.circulation.infrastructure.storage.AutomatedPatronBlocksRepository;
 import org.folio.circulation.infrastructure.storage.ConfigurationRepository;
-import org.folio.circulation.infrastructure.storage.RequestRepository;
+import org.folio.circulation.infrastructure.storage.requests.RequestRepository;
 
 public class CreateRequestRepositories {
   private final RequestRepository requestRepository;
