+++ src/main/java/org/folio/circulation/domain/validation/ClosedRequestValidator.java
@@ -7,7 +7,7 @@ import java.util.concurrent.CompletableFuture;
 
 import org.folio.circulation.domain.Request;
 import org.folio.circulation.domain.RequestAndRelatedRecords;
-import org.folio.circulation.infrastructure.storage.RequestRepository;
+import org.folio.circulation.infrastructure.storage.requests.RequestRepository;
 import org.folio.circulation.support.Result;
 
 public class ClosedRequestValidator {
